requirements.txt
Flask==2.2.5
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
tensorflow==2.13.0
opencv-python==4.8.0.76
matplotlib==3.7.1
Pillow==9.5.0